Adding `debounce` API
This is alternate way to invoke `clear` or `rebuild` that puts the call in a queue that gets processed at the end of the Laravel request.  The queue is deduped as you add to it.  Thus, you can have event listeners call (for example) `Freezer::rebuild()` many times, but only rebuild your cache once.  Closes #16.